Refactoring Step 4: Split out a new Bubble class from Comical
